// Complementarios/Complementario 3 Parcial 2/Client-Web/src/Interfaces/IBuses.ts
// Generated by https://quicktype.io

export interface IResBus {
    total: number;
    datos: Bus[];
}

export interface Bus {
    _id?:           string;
    Placa:         string;
    NumeroBus:     string;
    Capacidad:     number;
    CedChofer:     string;
    NombresChofer: string;
    Estado?:       boolean;
    _v?:           number;
    Contraseña:    string;
}